refactor: replace SmartStickyButton with PageControlButtons and remove unused scroll messages
