+++ front/app/components/Form/Components/Layouts/PageControlButtons.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 
 import {
   Box,
@@ -70,6 +70,8 @@ interface Props {
   pageButtonLabelMultiloc?: Multiloc;
   pageButtonLink?: string;
   project: IProject | undefined;
+  pageQuestions?: Array<{ id: string; key: string }>;
+  currentPageNumber?: number;
 }
 
 const PageControlButtons = ({
@@ -83,6 +85,8 @@ const PageControlButtons = ({
   pageButtonLink,
   currentPhase,
   project,
+  pageQuestions = [],
+  currentPageNumber,
 }: Props) => {
   const theme = useTheme();
   const localize = useLocalize();
@@ -92,6 +96,57 @@ const PageControlButtons = ({
   const userCanModerate = project
     ? canModerateProject(project.data, authUser)
     : false;
+
+  const [hasReachedBottom, setHasReachedBottom] = useState(false);
+  const observerRef = useRef<IntersectionObserver | null>(null);
+
+  useEffect(() => {
+    setHasReachedBottom(false);
+  }, [currentPageNumber]);
+
+  useEffect(() => {
+    const enabledQuestions = pageQuestions.filter((q) => q.id && q.key);
+    const lastQuestion =
+      enabledQuestions.length > 0
+        ? enabledQuestions[enabledQuestions.length - 1]
+        : undefined;
+
+    if (!lastQuestion) {
+      setHasReachedBottom(true);
+      return;
+    }
+
+    const lastQuestionElement = document.querySelector(
+      `[data-question-id="${lastQuestion.id}"]`
+    );
+
+    if (!lastQuestionElement) {
+      setHasReachedBottom(true);
+      return;
+    }
+
+    observerRef.current = new IntersectionObserver(
+      (entries) => {
+        entries.forEach((entry) => {
+          if (entry.isIntersecting) {
+            setHasReachedBottom(true);
+          }
+        });
+      },
+      {
+        threshold: 0,
+        rootMargin: '0px 0px 0px 0px',
+      }
+    );
+
+    observerRef.current.observe(lastQuestionElement);
+
+    return () => {
+      if (observerRef.current) {
+        observerRef.current.disconnect();
+      }
+    };
+  }, [pageQuestions, currentPageNumber]);
 
   const getButtonMessage = () => {
     if (pageVariant !== 'after-submission') {
@@ -121,6 +176,9 @@ const PageControlButtons = ({
     const inputTerm = getInputTerm(phases, currentPhase);
     return formatMessage(inputTermMessages[inputTerm]);
   };
+
+  const isButtonDisabled =
+    !hasReachedBottom && pageVariant !== 'after-submission';
 
   return (
     <Box
@@ -166,6 +224,7 @@ const PageControlButtons = ({
             boxShadow={defaultStyles.boxShadow}
             processing={isLoading}
             linkTo={pageButtonLink}
+            disabled={isButtonDisabled}
           >
             {getButtonMessage()}
           </ButtonWithLink>
@@ -178,6 +237,7 @@ const PageControlButtons = ({
             bgColor={theme.colors.tenantPrimary}
             boxShadow={defaultStyles.boxShadow}
             processing={isLoading}
+            disabled={isButtonDisabled}
           >
             {getButtonMessage()}
           </Button>
